# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://car-project-io.test"
        }
    ],
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "List all cars",
                "description": "Retrieve a paginated list of cars with optional filtering, sorting, and including related data",
                "operationId": "158b081c3b62cbad1d47a7baf9ed52ac",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field (id, make, year)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "make",
                                "year"
                            ]
                        }
                    },
                    {
                        "name": "filter[make]",
                        "in": "query",
                        "description": "Filter by car make",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[model]",
                        "in": "query",
                        "description": "Filter by car model",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[year]",
                        "in": "query",
                        "description": "Filter by car year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[nickname]",
                        "in": "query",
                        "description": "Filter by car nickname",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[vin]",
                        "in": "query",
                        "description": "Filter by VIN",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "include[]",
                        "in": "query",
                        "description": "Include related data (user, modifications). Can be used multiple times: include[]=user&include[]=modifications",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "user",
                                    "modifications"
                                ]
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CarResource"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create a new car",
                "description": "Create a new car for the authenticated user",
                "operationId": "581b89650bc67a7ae22937830482fcb7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "make",
                                    "model",
                                    "year"
                                ],
                                "properties": {
                                    "make": {
                                        "description": "Car manufacturer",
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "model": {
                                        "description": "Car model",
                                        "type": "string",
                                        "example": "Camry"
                                    },
                                    "year": {
                                        "description": "Manufacturing year",
                                        "type": "integer",
                                        "example": 2020
                                    },
                                    "nickname": {
                                        "description": "Optional nickname for the car",
                                        "type": "string",
                                        "example": "My Daily Driver"
                                    },
                                    "vin": {
                                        "description": "Vehicle Identification Number",
                                        "type": "string",
                                        "example": "1HGBH41JXMN109186"
                                    },
                                    "notes": {
                                        "description": "Additional notes about the car",
                                        "type": "string",
                                        "example": "Great condition, low mileage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Car created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CarResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CarResource": {
                "properties": {
                    "id": {
                        "description": "Car ID",
                        "type": "integer",
                        "example": 1
                    },
                    "make": {
                        "description": "Car manufacturer",
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "description": "Car model",
                        "type": "string",
                        "example": "Camry"
                    },
                    "year": {
                        "description": "Manufacturing year",
                        "type": "integer",
                        "example": 2020
                    },
                    "nickname": {
                        "description": "Optional nickname for the car",
                        "type": "string",
                        "example": "My Daily Driver",
                        "nullable": true
                    },
                    "vin": {
                        "description": "Vehicle Identification Number",
                        "type": "string",
                        "example": "1HGBH41JXMN109186",
                        "nullable": true
                    },
                    "image_url": {
                        "description": "URL to car image",
                        "type": "string",
                        "example": "https://example.com/car.jpg",
                        "nullable": true
                    },
                    "notes": {
                        "description": "Additional notes about the car",
                        "type": "string",
                        "example": "Great condition, low mileage",
                        "nullable": true
                    },
                    "user": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/UserResource"
                            }
                        ],
                        "nullable": true
                    },
                    "modifications": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ModificationResource"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "john@example.com"
                    }
                },
                "type": "object"
            },
            "ModificationResource": {
                "properties": {
                    "id": {
                        "description": "Modification ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Modification name",
                        "type": "string",
                        "example": "Performance Exhaust"
                    },
                    "description": {
                        "description": "Modification description",
                        "type": "string",
                        "example": "High-performance exhaust system",
                        "nullable": true
                    },
                    "cost": {
                        "description": "Modification cost",
                        "type": "number",
                        "format": "float",
                        "example": 599.99,
                        "nullable": true
                    },
                    "installed_at": {
                        "description": "Installation date",
                        "type": "string",
                        "format": "date",
                        "example": "2023-06-15",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "description": "Link to first page",
                        "type": "string",
                        "example": "/api/cars?page=1"
                    },
                    "last": {
                        "description": "Link to last page",
                        "type": "string",
                        "example": "/api/cars?page=10"
                    },
                    "prev": {
                        "description": "Link to previous page",
                        "type": "string",
                        "example": "/api/cars?page=1",
                        "nullable": true
                    },
                    "next": {
                        "description": "Link to next page",
                        "type": "string",
                        "example": "/api/cars?page=3",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer",
                        "example": 2
                    },
                    "from": {
                        "description": "First item number on current page",
                        "type": "integer",
                        "example": 11
                    },
                    "last_page": {
                        "description": "Last page number",
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "description": "Base path for pagination links",
                        "type": "string",
                        "example": "/api/cars"
                    },
                    "per_page": {
                        "description": "Items per page",
                        "type": "integer",
                        "example": 10
                    },
                    "to": {
                        "description": "Last item number on current page",
                        "type": "integer",
                        "example": 20
                    },
                    "total": {
                        "description": "Total number of items",
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}